add formatter mixin
